+++ src/android/FirebasePluginMessagingService.java
@@ -6,6 +6,8 @@ import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.BitmapFactory;
+import android.graphics.Bitmap;
+
 import android.media.RingtoneManager;
 import android.net.Uri;
 import android.os.Build;
@@ -14,13 +16,25 @@ import androidx.core.app.NotificationCompat;
 import android.util.Log;
 import android.app.Notification;
 import android.text.TextUtils;
+import android.text.Html;
+import android.text.Spanned;
 import android.content.ContentResolver;
 import android.graphics.Color;
+import android.graphics.PorterDuff;
+import android.graphics.PorterDuffXfermode;
+import android.graphics.Rect;
+import android.graphics.RectF;
+import android.graphics.Paint;
+import android.graphics.Canvas;
 
 import com.google.firebase.crashlytics.FirebaseCrashlytics;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.HttpURLConnection;
+import java.net.URL;
 import java.util.Map;
 import java.util.Random;
 
@@ -31,6 +45,8 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
     static final String defaultSmallIconName = "notification_icon";
     static final String defaultLargeIconName = "notification_icon_large";
 
+    static final String imageTypeCircle = "circle";
+    static final String imageTypeBigPicture = "big_picture";
 
     /**
      * Called if InstanceID token is updated. This may occur if the security of
@@ -48,6 +64,20 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
         }
     }
 
+    public Bitmap getBitmapFromURL(String strURL) {
+        try {
+            URL url = new URL(strURL);
+            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+            connection.setConnectTimeout(15000);
+            connection.setDoInput(true);
+            connection.connect();
+            InputStream input = connection.getInputStream();
+            return BitmapFactory.decodeStream(input);
+        } catch (IOException e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
 
     /**
      * Called when message is received.
@@ -89,6 +119,7 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
             String messageType;
             String title = null;
             String body = null;
+            String bodyHtml = null;
             String id = null;
             String sound = null;
             String vibrate = null;
@@ -98,6 +129,8 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
             String channelId = null;
             String visibility = null;
             String priority = null;
+            String image = null;
+            String imageType = null;
             boolean foregroundNotification = false;
 
             Map<String, String> data = remoteMessage.getData();
@@ -116,6 +149,9 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
                 sound = notification.getSound();
                 color = notification.getColor();
                 icon = notification.getIcon();
+                if (notification.getImageUrl() != null) {
+                    image = notification.getImageUrl().toString();
+                }
             }else{
                 Log.i(TAG, "Received message: data");
                 messageType = "data";
@@ -128,6 +164,7 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
                 }
                 if(data.containsKey("notification_title")) title = data.get("notification_title");
                 if(data.containsKey("notification_body")) body = data.get("notification_body");
+                if(data.containsKey("notification_android_body_html")) bodyHtml = data.get("notification_android_body_html");
                 if(data.containsKey("notification_android_channel_id")) channelId = data.get("notification_android_channel_id");
                 if(data.containsKey("notification_android_id")) id = data.get("notification_android_id");
                 if(data.containsKey("notification_android_sound")) sound = data.get("notification_android_sound");
@@ -137,6 +174,8 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
                 if(data.containsKey("notification_android_icon")) icon = data.get("notification_android_icon");
                 if(data.containsKey("notification_android_visibility")) visibility = data.get("notification_android_visibility");
                 if(data.containsKey("notification_android_priority")) priority = data.get("notification_android_priority");
+                if(data.containsKey("notification_android_image")) image = data.get("notification_android_image");
+                if(data.containsKey("notification_android_image_type")) imageType = data.get("notification_android_image_type");
             }
 
             if (TextUtils.isEmpty(id)) {
@@ -157,18 +196,20 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
             Log.d(TAG, "Channel Id: " + channelId);
             Log.d(TAG, "Visibility: " + visibility);
             Log.d(TAG, "Priority: " + priority);
+            Log.d(TAG, "Image: " + image);
+            Log.d(TAG, "image Type: " + imageType);
 
 
             if (!TextUtils.isEmpty(body) || !TextUtils.isEmpty(title) || (data != null && !data.isEmpty())) {
                 boolean showNotification = (FirebasePlugin.inBackground() || !FirebasePlugin.hasNotificationsCallback() || foregroundNotification) && (!TextUtils.isEmpty(body) || !TextUtils.isEmpty(title));
-                sendMessage(remoteMessage, data, messageType, id, title, body, showNotification, sound, vibrate, light, color, icon, channelId, priority, visibility);
+                sendMessage(remoteMessage, data, messageType, id, title, body, bodyHtml, showNotification, sound, vibrate, light, color, icon, channelId, priority, visibility, image, imageType);
             }
         }catch (Exception e){
             FirebasePlugin.handleExceptionWithoutContext(e);
         }
     }
 
-    private void sendMessage(RemoteMessage remoteMessage, Map<String, String> data, String messageType, String id, String title, String body, boolean showNotification, String sound, String vibrate, String light, String color, String icon, String channelId, String priority, String visibility) {
+    private void sendMessage(RemoteMessage remoteMessage, Map<String, String> data, String messageType, String id, String title, String body, String bodyHtml, boolean showNotification, String sound, String vibrate, String light, String color, String icon, String channelId, String priority, String visibility, String image, String imageType) {
         Log.d(TAG, "sendMessage(): messageType="+messageType+"; showNotification="+showNotification+"; id="+id+"; title="+title+"; body="+body+"; sound="+sound+"; vibrate="+vibrate+"; light="+light+"; color="+color+"; icon="+icon+"; channel="+channelId+"; data="+data.toString());
         Bundle bundle = new Bundle();
         for (String key : data.keySet()) {
@@ -178,6 +219,7 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
         this.putKVInBundle("id", id, bundle);
         this.putKVInBundle("title", title, bundle);
         this.putKVInBundle("body", body, bundle);
+        this.putKVInBundle("body_html", bodyHtml, bundle);
         this.putKVInBundle("sound", sound, bundle);
         this.putKVInBundle("vibrate", vibrate, bundle);
         this.putKVInBundle("light", light, bundle);
@@ -186,6 +228,8 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
         this.putKVInBundle("channel_id", channelId, bundle);
         this.putKVInBundle("priority", priority, bundle);
         this.putKVInBundle("visibility", visibility, bundle);
+        this.putKVInBundle("image", image, bundle);
+        this.putKVInBundle("image_type", imageType, bundle);
         this.putKVInBundle("show_notification", String.valueOf(showNotification), bundle);
         this.putKVInBundle("from", remoteMessage.getFrom(), bundle);
         this.putKVInBundle("collapse_key", remoteMessage.getCollapseKey(), bundle);
@@ -205,13 +249,24 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
                 Log.d(TAG, "Channel ID: "+channelId);
             }
 
+
+
             NotificationCompat.Builder notificationBuilder = new NotificationCompat.Builder(this, channelId);
             notificationBuilder
                     .setContentTitle(title)
-                    .setContentText(body)
-                    .setStyle(new NotificationCompat.BigTextStyle().bigText(body))
                     .setAutoCancel(true)
                     .setContentIntent(pendingIntent);
+
+            if(bodyHtml != null) {
+                notificationBuilder
+                    .setContentText(fromHtml(body))
+                    .setStyle(new NotificationCompat.BigTextStyle().bigText(fromHtml(body)));
+            }else{
+                notificationBuilder
+                    .setContentText(body)
+                    .setStyle(new NotificationCompat.BigTextStyle().bigText(body));
+            }
+
 
             // On Android O+ the sound/lights/vibration are determined by the channel ID
             if(Build.VERSION.SDK_INT < Build.VERSION_CODES.O){
@@ -299,6 +354,21 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
                 }
             }
 
+            // Image
+            if (image != null) {
+                Log.d(TAG, "Large icon: image="+image);
+                Bitmap bitmap = getBitmapFromURL(image);
+                if(bitmap != null) {
+                    if(imageTypeCircle.equalsIgnoreCase(imageType)) {
+                        bitmap = getCircleBitmap(bitmap);
+                    }
+                    else if(imageTypeBigPicture.equalsIgnoreCase(imageType)) {
+                        notificationBuilder.setStyle(new NotificationCompat.BigPictureStyle().bigPicture(bitmap).bigLargeIcon(null));
+                    }
+                    notificationBuilder.setLargeIcon(bitmap);
+                }
+            }
+
             // Color
             if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.M) {
                 int defaultColor = getResources().getColor(getResources().getIdentifier("accent", "color", getPackageName()), null);
@@ -338,6 +408,42 @@ public class FirebasePluginMessagingService extends FirebaseMessagingService {
         }
         // Send to plugin
         FirebasePlugin.sendMessage(bundle, this.getApplicationContext());
+    }
+
+    private Bitmap getCircleBitmap(Bitmap bitmap) {
+
+        if (bitmap == null) {
+            return null;
+        }
+
+        final Bitmap output = Bitmap.createBitmap(bitmap.getWidth(), bitmap.getHeight(), Bitmap.Config.ARGB_8888);
+        final Canvas canvas = new Canvas(output);
+        final int color = Color.RED;
+        final Paint paint = new Paint();
+        final Rect rect = new Rect(0, 0, bitmap.getWidth(), bitmap.getHeight());
+        final RectF rectF = new RectF(rect);
+
+        paint.setAntiAlias(true);
+        canvas.drawARGB(0, 0, 0, 0);
+        paint.setColor(color);
+        float cx = bitmap.getWidth() / 2;
+        float cy = bitmap.getHeight() / 2;
+        float radius = cx < cy ? cx : cy;
+        canvas.drawCircle(cx, cy, radius, paint);
+
+        paint.setXfermode(new PorterDuffXfermode(PorterDuff.Mode.SRC_IN));
+        canvas.drawBitmap(bitmap, rect, rect, paint);
+
+        bitmap.recycle();
+
+        return output;
+    }
+
+    private Spanned fromHtml(String source) {
+        if (source != null)
+            return Html.fromHtml(source);
+        else
+            return null;
     }
 
     private void putKVInBundle(String k, String v, Bundle b){